Fix issue #2
This commit adds the file hyperram_errata.vhd that provides a
work-around for the errata mentioned in this issue.

The work-around may be disabled by setting the generic
G_ERRATA_ISSI_D_FIX to false (it is default enabled).

This commit also adds some refactoring of the Example Design. The test
parameters are controlled by the constants in top.vhd.